refactor: extract generate api key stage into component
